+++ src/regional_economy/entities/__init__.py
@@ -2,14 +2,17 @@
 
 from .business import Business, Sector
 from .government import Government
+from .healthcare import AgeCohort, HealthcareSystem
 from .household import Household, HouseholdAllocation
 from .region import Region
 from .university import StudentCohort, University
 from .visitor import TourismBusiness, TourismSector, Visitor
 
 __all__ = [
+    "AgeCohort",
     "Business",
     "Government",
+    "HealthcareSystem",
     "Household",
     "HouseholdAllocation",
     "Region",